Feat: 번역본 상태 변경 기능 구현

+++ backend/docshund/src/test/java/com/ssafy/docshund/domain/docs/service/DocsServiceImplTest.java
@@ -1,14 +1,10 @@
 package com.ssafy.docshund.domain.docs.service;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Objects;
 
-import com.ssafy.docshund.domain.docs.entity.*;
-import com.ssafy.docshund.domain.docs.repository.*;
-import com.ssafy.docshund.domain.users.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -16,15 +12,28 @@ import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cache.CacheManager;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.ssafy.docshund.domain.docs.dto.DocumentDto;
 import com.ssafy.docshund.domain.docs.dto.OriginDocumentDto;
 import com.ssafy.docshund.domain.docs.dto.TranslatedDocumentDto;
+import com.ssafy.docshund.domain.docs.entity.Document;
+import com.ssafy.docshund.domain.docs.entity.DocumentLike;
+import com.ssafy.docshund.domain.docs.entity.OriginDocument;
+import com.ssafy.docshund.domain.docs.entity.Position;
+import com.ssafy.docshund.domain.docs.entity.Status;
+import com.ssafy.docshund.domain.docs.entity.TranslatedDocument;
+import com.ssafy.docshund.domain.docs.entity.TranslatedDocumentLike;
+import com.ssafy.docshund.domain.docs.repository.CustomDocumentRepository;
+import com.ssafy.docshund.domain.docs.repository.DocumentLikeRepository;
+import com.ssafy.docshund.domain.docs.repository.DocumentRepository;
+import com.ssafy.docshund.domain.docs.repository.OriginDocumentRepository;
+import com.ssafy.docshund.domain.docs.repository.TranslatedDocumentLikeRepository;
+import com.ssafy.docshund.domain.docs.repository.TranslatedDocumentRepository;
 import com.ssafy.docshund.domain.users.entity.Provider;
 import com.ssafy.docshund.domain.users.entity.User;
+import com.ssafy.docshund.domain.users.repository.UserRepository;
 import com.ssafy.docshund.fixture.UserTestHelper;
 import com.ssafy.docshund.fixture.WithMockCustomOAuth2User;
 import com.ssafy.docshund.global.util.user.UserUtil;
@@ -59,7 +68,7 @@ class DocsServiceImplTest {
 	@Autowired
 	private CustomDocumentRepository customDocumentRepository;
 
-	private User user1;	// 어드민 유저
+	private User user1;    // 어드민 유저
 	private User user2; // 좋아요 및 번역 CUD 테스트용 유저
 	private User user3; // 좋아요 및 번역 CUD 테스트용 유저
 
@@ -76,12 +85,15 @@ class DocsServiceImplTest {
 	@BeforeEach
 	public void setUp() {
 		// user repository setup
-		user1 = userTestHelper.saveUser("admin@gmail.com", "100000", "adminUser", Provider.GOOGLE, true, "안녕하세요", "Backend",
-				true);
-		user2 = userTestHelper.saveUser("test1@gmail.com", "10001", "testUser1", Provider.GOOGLE, false, "안녕하세요", "Frontend",
-				true);
-		user3 = userTestHelper.saveUser("test2@github.com", "10002", "testUser2", Provider.GITHUB, false, "안녕하세요", "Backend",
-				true);
+		user1 = userTestHelper.saveUser("admin@gmail.com", "100000", "adminUser", Provider.GOOGLE, true, "안녕하세요",
+			"Backend",
+			true);
+		user2 = userTestHelper.saveUser("test1@gmail.com", "10001", "testUser1", Provider.GOOGLE, false, "안녕하세요",
+			"Frontend",
+			true);
+		user3 = userTestHelper.saveUser("test2@github.com", "10002", "testUser2", Provider.GITHUB, false, "안녕하세요",
+			"Backend",
+			true);
 		// document repository setup
 		doc1 = documentRepository.save(new Document("Spring", "Spring Boot", "logoImage", "v1",
 			0, Position.BACKEND, "apache", "docLink"));
@@ -94,12 +106,16 @@ class DocsServiceImplTest {
 		documentLikeRepository.save(new DocumentLike(doc2, user2));
 
 		// origin document repository setup
-		originDoc1 = originDocumentRepository.save(new OriginDocument(doc1, 1, "<p>", "<p>Spring Framework is the best Framework for Java.</p>"));
-		originDoc2 = originDocumentRepository.save(new OriginDocument(doc1, 2, "<p>", "<p>And Spring boot is very easy and simple.</p>"));
+		originDoc1 = originDocumentRepository.save(
+			new OriginDocument(doc1, 1, "<p>", "<p>Spring Framework is the best Framework for Java.</p>"));
+		originDoc2 = originDocumentRepository.save(
+			new OriginDocument(doc1, 2, "<p>", "<p>And Spring boot is very easy and simple.</p>"));
 
 		// translated document repository setup
-		transDoc1 = translatedDocumentRepository.save(new TranslatedDocument(originDoc1, user2, "스프링 프레임워크는 자바를 위한 최고의 프레임워크입니다.", 0, Status.VISIBLE));
-		transDoc2 = translatedDocumentRepository.save(new TranslatedDocument(originDoc2, user3, "그리고 스프링 부트는 아주 쉽고 간단합니다.", 0, Status.VISIBLE));
+		transDoc1 = translatedDocumentRepository.save(
+			new TranslatedDocument(originDoc1, user2, "스프링 프레임워크는 자바를 위한 최고의 프레임워크입니다.", 0, Status.VISIBLE));
+		transDoc2 = translatedDocumentRepository.save(
+			new TranslatedDocument(originDoc2, user3, "그리고 스프링 부트는 아주 쉽고 간단합니다.", 0, Status.VISIBLE));
 
 		// translated document like repository setup
 		translatedDocumentLikeRepository.save(new TranslatedDocumentLike(transDoc1, user2));
@@ -255,7 +271,8 @@ class DocsServiceImplTest {
 		String content = "새 번역 내용";
 
 		// when
-		TranslatedDocumentDto result = docsService.createTranslatedDocument(doc1.getDocsId(), originDoc2.getOriginId(), user2, content);
+		TranslatedDocumentDto result = docsService.createTranslatedDocument(doc1.getDocsId(), originDoc2.getOriginId(),
+			user2, content);
 
 		// then
 		assertThat(result).isNotNull();
@@ -269,7 +286,8 @@ class DocsServiceImplTest {
 		String content = "수정된 번역 내용";
 
 		// when
-		TranslatedDocumentDto result = docsService.updateTranslatedDocument(doc1.getDocsId(), Math.toIntExact(transDoc1.getTransId()), user2, content);
+		TranslatedDocumentDto result = docsService.updateTranslatedDocument(doc1.getDocsId(),
+			(long)Math.toIntExact(transDoc1.getTransId()), user2, content);
 
 		// then
 		assertThat(result).isNotNull();
@@ -282,7 +300,7 @@ class DocsServiceImplTest {
 	void deleteTranslatedDocument() {
 
 		// when
-		docsService.deleteTranslatedDocument(doc1.getDocsId(), Math.toIntExact(transDoc1.getTransId()), user2);
+		docsService.deleteTranslatedDocument(doc1.getDocsId(), (long)Math.toIntExact(transDoc1.getTransId()), user2);
 
 		// then
 		log.info("번역 삭제 성공");
@@ -294,7 +312,8 @@ class DocsServiceImplTest {
 	void toggleVotes() {
 
 		// when
-		boolean result = docsService.toggleVotes(doc1.getDocsId(), Math.toIntExact(transDoc1.getTransId()), user1);
+		boolean result = docsService.toggleVotes(doc1.getDocsId(), (long)Math.toIntExact(transDoc1.getTransId()),
+			user1);
 
 		// then
 		assertThat(result).isTrue();
